\name{markowitz}
\docType{data}
\alias{markowitz}
\title{markowitz 1959 data}
\description{A copy of the securities data, used in Markowitz's 1959 book, more specifically a multivariate timeseries, representing returns on nine securities over 18 years.}
\format{A table with 18 realisations of 10 variables (1 time variable and 9 response variables).}

