\name{tcp}
\alias{thm}
\alias{tcp}
\title{triangular heatmaps and triangular contour plots}
\description{The thm function is an early prototype for triangular heat maps (using vectorised graphics). The tcp function is intended to produce triangular contour plots, however currently just calls thm.}
\usage{
thm (f, n=25, \dots)
tcp (f, \dots)
}
\arguments{
\item{f}{A pr object.}
\item{n}{Resolution of the heatmap (n points per side).}
\item{\dots}{.}
}


