\name{tcp}
\alias{tcp}
\alias{thm}
\title{triangular contour plots and triangular heatmaps}
\description{The thm function is an early prototype for triangular heat maps (using vectorised graphics). The tcp function is intended to produce triangular contour plots, however currently just calls thm.}
\usage{
tcp (f, \dots)
thm (f, n=17, \dots)
}
\arguments{
\item{f}{A pr object.}
\item{n}{Resolution of the heatmap (n points per side).}
\item{\dots}{.}
}

